# docs/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "Skill Assessment Service:",
    "version": "1.0.0"
  },
  "host": "localhost:8084",
  "basePath": "/",
  "paths": {
...
    "/api/user": {
      "post": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          },
          {
            "securityDefinitions": []
          },
          {
            "Bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          },
          {
            "requests": []
          }
        ],
        "description": "Create new user",
        "tags": [
          "admin"
        ],
        "operationId": "UserRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/SuccessResponse"
          },
          "400": {
            "$ref": "#/responses/BadRequestErrorResponse"
          },
          "401": {
            "$ref": "#/responses/UnAuthenticatedAccessResponse"
          },
          "403": {
            "$ref": "#/responses/UnAuthorizedAccessResponse"
          },
          "409": {
            "$ref": "#/responses/ConflictErrorResponse"
          },
          "500": {
            "$ref": "#/responses/InternalServerErrorResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
    "BadRequestError": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message",
          "example": "Invalid parameters"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status",
          "example": 400
        },
        "timestamp": {
          "type": "string",
          "x-go-name": "TimeStamp",
          "example": "2022-05-20 16:59:05"
        }
      },
      "x-go-package": "github.com/sandy0786/skill-assessment-service/errors"
    },
...
    "ConflictError": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message",
          "example": "Already exist"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status",
          "example": 409
        },
        "timestamp": {
          "type": "string",
          "x-go-name": "TimeStamp",
          "example": "2022-05-20 16:59:05"
        }
      },
      "x-go-package": "github.com/sandy0786/skill-assessment-service/errors"
    },
    "GlobalError": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message",
          "example": "Something went wrong"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status",
          "example": 500
        },
        "timestamp": {
          "type": "string",
          "x-go-name": "TimeStamp",
          "example": "2022-05-20 16:59:05"
        }
      },
      "x-go-package": "github.com/sandy0786/skill-assessment-service/errors"
    },
...
    "SuccessResponse": {
      "description": "Success response",
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message",
          "example": "Data saved successfully"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status",
          "example": 200
        },
        "timestamp": {
          "type": "string",
          "x-go-name": "TimeStamp",
          "example": "2022-05-20 16:59:05"
        }
      },
      "x-go-package": "github.com/sandy0786/skill-assessment-service/response/success"
    },
    "UnAuthenticatedAccessError": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message",
          "example": "User is not authenticated"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status",
          "example": 401
        },
        "timestamp": {
          "type": "string",
          "x-go-name": "TimeStamp",
          "example": "2022-05-20 16:59:05"
        }
      },
      "x-go-package": "github.com/sandy0786/skill-assessment-service/errors"
    },
    "UnAuthorizedAccessError": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message",
          "example": "User is not authorized"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status",
          "example": 401
        },
        "timestamp": {
          "type": "string",
          "x-go-name": "TimeStamp",
          "example": "2022-05-20 16:59:05"
        }
      },
      "x-go-package": "github.com/sandy0786/skill-assessment-service/errors"
    },
...
    "UserRequest": {
      "description": "User Request",
      "type": "object",
      "required": [
        "username",
        "password",
        "email",
        "roleId"
      ],
      "properties": {
        "email": {
          "description": "the email for this user",
          "type": "string",
          "x-go-name": "Email",
          "example": "user@provider.com"
        },
        "password": {
          "description": "the password for this user",
          "type": "string",
          "minLength": 8,
          "x-go-name": "Password"
        },
        "roleId": {
          "description": "the roleId for this user",
          "type": "string",
          "x-go-name": "Role",
          "example": "62d64f2142dac7953ac4ff32"
        },
        "username": {
          "description": "the username for this user",
          "type": "string",
          "x-go-name": "Username"
        }
      },
      "x-go-package": "github.com/sandy0786/skill-assessment-service/request/user"
    },
...
  },
  "responses": {
    "BadRequestErrorResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/BadRequestError"
      }
    },
...
    "ConflictErrorResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/ConflictError"
      }
    },
...
    "InternalServerErrorResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/GlobalError"
      }
    },
...
    "SuccessResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/SuccessResponse"
      }
    },
    "UnAuthenticatedAccessResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/UnAuthenticatedAccessError"
      }
    },
    "UnAuthorizedAccessResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/UnAuthorizedAccessError"
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "Bearer": [
        "[]"
      ]
    }
  ]
}